api/requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
fastapi>=0.100.0
uvicorn>=0.22.0
pytest>=7.0.0
